move modules into model concerns and name them
The system test only has to change a little, but now the presence of
several different modules in the system test indicates an opportunity to
abstract away and encapsulate making room for more ideas.

You could even build another module that takes the input file and
validates all of the unique lookup list inputs against the ServiceNow
database, so that you can pre-empt the production load and make sure
in advance that our input file listing is ready for production use.