more attention to optimizing across runs
Store the Business Application's URL so we do not need to search for it
the second time that it is updated!

Also, more output, and some commented out binding.pry statements
wherever you might want to take a peek... and some additional attention
toward the consistent and correct formatting of the output.